requirements.txt

editdistance==0.8.1
Jinja2==3.1.4
joblib==1.4.2
loguru==0.7.2
matplotlib==3.10.1
numpy==2.2.5
pandas==2.2.3
plotly==5.22.0
polars==0.20.31
scanpy==1.11.1
scikit-learn==1.5.2
scipy==1.15.2
seaborn==0.13.2
tqdm==4.67.1
python-igraph==0.11.6
ahocorasick-python==0.0.9
anndata==0.11.4
biopython==1.85
colorama==0.4.6
cutadapt==5.1
dnaio==1.2.3
opencv-python==4.12.0.88
pillow==11.2.1
playwright==1.53.0
pysam==0.23.3
python-docx==1.2.0
typer==0.16.0
typing_extensions==4.13.2
